Fix readme (#31)
* Get the git tree to the correct path

Changed the "obtaining the code" section to use git clone as go get
won't get a tree when the root of that tree has no "*.go" files.

* Use depend-update in case the GOPATH has old dependencies

If the GOPATH has old dependencies, "Solving failures" may happen as
those old dependencies will try to be used and may not have the needed
functions. Use the latest dependencies to prevent this problem.